[ONEM-11764] Submit secondary email addresses to SQ as SCM accounts
The original sonar-auth-github plugin fetches the primary email address
of the user being authenticated from GitHub and stores it as its email
address in SonarQube. Unfortunately some GitHub users use multiple
email addresses. Having them all in SonarQube listed as SCM accounts is
crutial to determine the assignee of a given issue by 'git blame' or
equivalent, when the issue is first reported to SonarQube.

SonarSource is already aware of this shortcoming [1][2], though it hasn't
been implemented yet as it requires changes in SonarQube's core and
plugin APIs.

[1] https://jira.sonarsource.com/browse/SONAR-10248
[2] https://jira.sonarsource.com/browse/SQAUTHGITH-18

This change is a workaround, far from being a good solution. It uses
SonarQube's Web API to submit secondary email addresses as user's SCM
accounts during the authentication process. It does that in a separate
thread to avoid influencing the authentication process in any way. One
of the drawbacks of this solution is that all SCM accounts already
stored in the user's account are lost and overwritten by secondary email
addresses. However in some cases this is acceptable.