[ONEM-11764] Force @libertyglobal.com emails as primary ones
An email address from libertyglobal.com domain is now always considered
primary regardless of what GitHub API says. This is because we want
emails from SonarQube to be routed to corporate email boxes rather than
private ones (which are often set as primary ones in GitHub, as GitHub
offers customizable email routing rules, which are not exposed via its
APIs).